MDTZ-1042: Refactor Jira inbound auth code
