+++ src/web/routes/lifecycle-events/integration.test.ts
@@ -22,7 +22,7 @@ import {
 	figmaTeamRepository,
 } from '../../../infrastructure/repositories';
 import {
-	generateInboundRequestAsymmetricJwtToken,
+	generateJiraAsymmetricJwtToken,
 	mockConnectGetKeyEndpoint,
 	mockFigmaDeleteWebhookEndpoint,
 	mockFigmaMeEndpoint,
@@ -36,16 +36,15 @@ describe('/lifecycleEvents', () => {
 				clientKey,
 			});
 			const keyId = uuidv4();
-			const { jwtToken, publicKey } =
-				await generateInboundRequestAsymmetricJwtToken({
-					keyId,
-					method: 'POST',
-					pathname: '/lifecycleEvents/installed',
-					connectInstallation: {
-						baseUrl: getConfig().app.baseUrl,
-						clientKey,
-					},
-				});
+			const { jwtToken, publicKey } = await generateJiraAsymmetricJwtToken({
+				keyId,
+				method: 'POST',
+				pathname: '/lifecycleEvents/installed',
+				connectInstallation: {
+					baseUrl: getConfig().app.baseUrl,
+					clientKey,
+				},
+			});
 
 			mockConnectGetKeyEndpoint({
 				baseUrl: getConfig().jira.connectKeyServerUrl,
@@ -73,16 +72,15 @@ describe('/lifecycleEvents', () => {
 
 		it('should respond 401 when JWT token is invalid (unknown issuer)', async () => {
 			const keyId = uuidv4();
-			const { jwtToken, publicKey } =
-				await generateInboundRequestAsymmetricJwtToken({
-					keyId,
-					method: 'POST',
-					pathname: '/incorrect-pathname',
-					connectInstallation: {
-						baseUrl: getConfig().app.baseUrl,
-						clientKey: uuidv4(),
-					},
-				});
+			const { jwtToken, publicKey } = await generateJiraAsymmetricJwtToken({
+				keyId,
+				method: 'POST',
+				pathname: '/incorrect-pathname',
+				connectInstallation: {
+					baseUrl: getConfig().app.baseUrl,
+					clientKey: uuidv4(),
+				},
+			});
 
 			mockConnectGetKeyEndpoint({
 				baseUrl: getConfig().jira.connectKeyServerUrl,
@@ -175,16 +173,15 @@ describe('/lifecycleEvents', () => {
 			]);
 
 			const keyId = uuidv4();
-			const { jwtToken, publicKey } =
-				await generateInboundRequestAsymmetricJwtToken({
-					keyId,
-					method: 'POST',
-					pathname: '/lifecycleEvents/uninstalled',
-					connectInstallation: {
-						baseUrl: getConfig().app.baseUrl,
-						clientKey: targetConnectInstallation.clientKey,
-					},
-				});
+			const { jwtToken, publicKey } = await generateJiraAsymmetricJwtToken({
+				keyId,
+				method: 'POST',
+				pathname: '/lifecycleEvents/uninstalled',
+				connectInstallation: {
+					baseUrl: getConfig().app.baseUrl,
+					clientKey: targetConnectInstallation.clientKey,
+				},
+			});
 
 			mockConnectGetKeyEndpoint({
 				baseUrl: getConfig().jira.connectKeyServerUrl,
@@ -230,16 +227,15 @@ describe('/lifecycleEvents', () => {
 
 		it('should respond 401 when JWT token is invalid (unknown issuer)', async () => {
 			const keyId = uuidv4();
-			const { jwtToken, publicKey } =
-				await generateInboundRequestAsymmetricJwtToken({
-					keyId,
-					method: 'POST',
-					pathname: '/incorrect-pathname',
-					connectInstallation: {
-						baseUrl: getConfig().app.baseUrl,
-						clientKey: uuidv4(),
-					},
-				});
+			const { jwtToken, publicKey } = await generateJiraAsymmetricJwtToken({
+				keyId,
+				method: 'POST',
+				pathname: '/incorrect-pathname',
+				connectInstallation: {
+					baseUrl: getConfig().app.baseUrl,
+					clientKey: uuidv4(),
+				},
+			});
 
 			mockConnectGetKeyEndpoint({
 				baseUrl: getConfig().jira.connectKeyServerUrl,
