+++ src/web/testing/index.ts
@@ -1,4 +1,4 @@
 export * from './connect-api-mock';
 export * from './figma-api-mocks';
 export * from './jira-api-mocks';
-export * from './mocks';
+export * from '../../infrastructure/jira/testing/mocks';
